Int becomes bitvector
Leon now matches Scala semantics of Int as 32 bits bit-vectors.
This commits modifies the semantics of IntLiteral to be treated
as 32 bits integer everywhere (solver, evaluator, ...).

Introduces a new literal type, InfiniteIntegerLiteral, representing
a natural integer. The front-end maps the use of BigInt to these
new trees, and the solver properly handles them as mathematical
integers.

The behaviour of many regression tests changes due to this new
semantics. In particular many of them now timeout because they
are no longer proving properties over mathematical integers. This
commit updates the tests to reflect this new semantics.

+++ valid/Anonymous.scala
@@ -1,9 +1,9 @@
 import leon.lang._
 
 object Anonymous {
-  def test(x: Int) = {
+  def test(x: BigInt) = {
     require(x > 0)
-    val i = (a: Int) => a + 1
+    val i = (a: BigInt) => a + 1
     i(x) + i(2)
   } ensuring { res => res > 0 }
 }
